apiCalls/requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
DateTime==4.2
Flask==0.12.2
Flask-Cors==3.0.3
get==0.0.21
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jsonify==0.5
MarkupSafe==1.0
numpy==1.13.0
pandas==0.20.2
post==0.0.13
psutil==5.2.2
psycopg2==2.7.1
public==0.0.38
python-dateutil==2.6.0
pytz==2017.2
query-string==0.0.12
request==0.0.13
requests==2.18.1
setupfiles==0.0.50
six==1.10.0
SQLAlchemy==1.1.11
urllib3==1.21.1
Werkzeug==0.12.2
zope.interface==4.4.2
